+++ apps/playground/CHANGELOG.md
@@ -1,5 +1,22 @@
 # @ai-gui/playground
 
+## 0.0.11
+
+### Patch Changes
+
+- Updated dependencies
+  - @ai-gui/core@0.7.0
+  - @ai-gui/devtools@0.7.0
+  - @ai-gui/plugin-artifact@0.7.0
+  - @ai-gui/plugin-citation@0.7.0
+  - @ai-gui/plugin-map@0.7.0
+  - @ai-gui/plugin-mermaid@0.7.0
+  - @ai-gui/plugin-molecule@0.7.0
+  - @ai-gui/plugin-ui@0.7.0
+  - @ai-gui/react@0.7.0
+  - @ai-gui/vanilla@0.7.0
+  - @ai-gui/vue@0.7.0
+
 ## 0.0.10
 
 ### Patch Changes